+++ lib/config/index.ts
