+++ lib/constants/init.ts
