+++ lib/helpers/fileHelpers.ts
