+++ lib/init/index.ts
